+++ AmazonChimeSDKDemo/AmazonChimeSDKDemoObjC/ViewControllerObjC.h
@@ -10,7 +10,7 @@
 #import <AVFoundation/AVFoundation.h>
 #import <AmazonChimeSDK/AmazonChimeSDK-Swift.h>
 
-@interface ViewControllerObjC : UIViewController <RealtimeObserver>
+@interface ViewControllerObjC : UIViewController <RealtimeObserver, MetricsObserver, VideoTileObserver>
 
 #define SERVER_URL "YOUR_SERVER_URL"
 #define SERVER_REGION "YOUR_SERVER_REGION"
@@ -21,6 +21,8 @@
 @property (weak, nonatomic) IBOutlet UILabel *versionLabel;
 @property (weak, nonatomic) IBOutlet UITextField *meetingIDText;
 @property (weak, nonatomic) IBOutlet UITextField *nameText;
+@property (weak, nonatomic) IBOutlet DefaultVideoRenderView *selfVideoView;
+@property (weak, nonatomic) IBOutlet DefaultVideoRenderView *remoteVideoView;
 
 @property (nonatomic, strong) ConsoleLogger *logger;
 @property (nonatomic, strong) DefaultMeetingSession *meetingSession;